+++ src/app/shared/card/card.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+/**
+ * Opakowanie dla innych komponentów
+ */
+@Component({
+  selector: 'app-card',
+  standalone: true,
+  imports: [],
+  templateUrl: './card.component.html',
+  styleUrl: './card.component.css'
+})
+export class CardComponent {
+}